requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
certifi==2025.8.3
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
mysqlclient==2.2.7
openai==1.102.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.10.0
